requirements.txt
openai-agents
chainlit
openai
python-dotenv
